Productos y Scripts (#2)
* Creacion de get productos

* Realizar rol y productos

* Scritps, modificación productos

+++ ArtInk/ArtInk.Application/Services/Interfaces/IServiceUsuario.cs
@@ -5,6 +5,6 @@ namespace ArtInk.Application.Services.Interfaces
     public interface IServiceUsuario
     {
         Task<ICollection<UsuarioDTO>> ListAsync();
-        Task<UsuarioDTO> FindByIdAsync(int id);
+        Task<UsuarioDTO> FindByIdAsync(short id);
     }
 }
